Laget felles vedtaksperiodeklasse. Trukket ut Utbetalingsperiode-utils og laget opphørsperiodeklasse.
