+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/vedtak/vedtaksperiode/Opphørsperiode.kt
@@ -1,0 +1,10 @@
+package no.nav.familie.ba.sak.behandling.vedtak.vedtaksperiode
+
+import java.time.LocalDate
+
+data class Opphørsperiode(
+        override val periodeFom: LocalDate,
+        override val periodeTom: LocalDate,
+        override val vedtaksperiodetype: Vedtaksperiodetype = Vedtaksperiodetype.OPPHØR
+) : Vedtaksperiode
+
